+++ source/extensions/omni.isaac.wheeled_robots/python/tests/test_differential_controller.py
@@ -89,6 +89,38 @@ class TestDifferentialControllerNode(ogts.OmniGraphTestCase):
         self.assertEqual(og.Controller(og.Controller.attribute("outputs:velocityCommand", diff_node)).get()[0], 8.125)
         self.assertEqual(og.Controller(og.Controller.attribute("outputs:velocityCommand", diff_node)).get()[1], 11.875)
 
+    async def test_differential_controller_node_acceleration_limits(self):
+        (test_diff_graph, [play_node, diff_node], _, _) = og.Controller.edit(
+            {"graph_path": "/ActionGraph", "evaluator_name": "execution"},
+            {
+                og.Controller.Keys.CREATE_NODES: [
+                    ("OnPlaybackTick", "omni.graph.action.OnPlaybackTick"),
+                    ("DifferentialController", "omni.isaac.wheeled_robots.DifferentialController"),
+                ],
+                og.Controller.Keys.CONNECT: [
+                    ("OnPlaybackTick.outputs:tick", "DifferentialController.inputs:execIn"),
+                ],
+                og.Controller.Keys.SET_VALUES: [
+                    ("DifferentialController.inputs:wheelRadius", 0.03),
+                    ("DifferentialController.inputs:wheelDistance", 0.1125),
+                    ("DifferentialController.inputs:linearVelocity", 0.3),
+                    ("DifferentialController.inputs:angularVelocity", 1.0),
+                    ("DifferentialController.inputs:maxAcceleration", 1.0),
+                    ("DifferentialController.inputs:maxDeceleration", 1.0),
+                    ("DifferentialController.inputs:dt", 1 / 60),
+                ],
+            },
+        )
+
+        self._timeline.play()
+        await omni.kit.app.get_app().next_update_async()
+        await simulate_async(0.05, 60)
+
+        # 8.125 and 11.875 are the expected velocity at full speed, with the acceleration limit, it should be less than those numbers
+        print(str(og.Controller(og.Controller.attribute("outputs:velocityCommand", diff_node)).get()))
+        self.assertLess(og.Controller(og.Controller.attribute("outputs:velocityCommand", diff_node)).get()[0], 8.125)
+        self.assertLess(og.Controller(og.Controller.attribute("outputs:velocityCommand", diff_node)).get()[1], 11.875)
+
     async def test_differential_controller_node_reset(self):
         (test_diff_graph, [play_node, diff_node], _, _) = og.Controller.edit(
             {"graph_path": "/ActionGraph", "evaluator_name": "execution"},